Remove check_metrics_vec_kernel_count from test_cpu_repro.py::CPUReproTests::test_transpose_non_contiguous (#141246)
The test was initially added due to accuracy issues which is sufficiently covered by the `self.common(fn, (x,))` assertion.

Unfortunately, the test fails due to tiling logic on `128-bit` vector size, which is outside the scope of this test and therefore it was overly specific.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/141246
Approved by: https://github.com/desertfire